+++ core-bundle/src/Resources/contao/config/config.php
@@ -226,7 +226,8 @@ $GLOBALS['BE_FFL'] = array
 	'keyValueWizard' => 'KeyValueWizard',
 	'imageSize'      => 'ImageSize',
 	'timePeriod'     => 'TimePeriod',
-	'metaWizard'     => 'MetaWizard'
+	'metaWizard'     => 'MetaWizard',
+	'sectionWizard'  => 'SectionWizard'
 );