+++ core-bundle/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### 4.3.0-beta1 (2016-XX-XX)
 
+ * Add more flexible custom layout sections (see contao/core#6630).
  * Make all wizards sortable via keyboard (see #583).
  * Add CSS classes to image sizes (see #555).
  * Improve the responsiveness of the back end theme.